pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "my description"
authors = [ {name = "My Name", email = "myemail@example.com"}, ]
requires-python = ">= 3.12"
dependencies = [
    "crewai[tools]>=0.108.0",
    "html2text>=2024.2.26",
    "python-dotenv>=1.1.0",
    "reflex>=0.7.5",
    "selenium>=4.31.0",
    "webdriver-manager>=4.0.2",
]

[tool.setuptools.packages.find]
where = ["morning_routine"]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
